+++ SignalBench.Core/Flux/Transport/ITransport.cs
@@ -1,0 +1,8 @@
+namespace SignalBench.Core.Flux.Transport;
+
+public interface ITransport : IDisposable
+{
+    void Send(byte[] data);
+    bool IsConnected { get; }
+    void Connect();
+}