+++ SignalBench.Core/Generation/SignalBase.cs
@@ -1,0 +1,6 @@
+namespace SignalBench.Core.Generation;
+
+public abstract class SignalBase
+{
+    public abstract double Evaluate(double time);
+}